+++ brownie/_cli/ethpm.py
@@ -8,7 +8,7 @@ import yaml
 from docopt import docopt
 
 from brownie import accounts, network
-from brownie._config import BROWNIE_FOLDER
+from brownie._config import BROWNIE_FOLDER, DATA_FOLDER
 from brownie.exceptions import ProjectNotFound, UnknownAccount
 from brownie.project import check_for_project, ethpm
 from brownie.utils import color, notify
@@ -148,7 +148,7 @@ def _release(project_path, registry_address, sender):
 
 
 def _all(project_path):
-    ethpm_folder = BROWNIE_FOLDER.joinpath("data/ethpm")
+    ethpm_folder = DATA_FOLDER.joinpath("ethpm")
     print(
         f"Visit {color('bright magenta')}https://explorer.ethpm.com/{color}"
         " for a list of ethPM registries and packages.\n"