+++ brownie/network/account.py
@@ -11,7 +11,7 @@ import eth_keys
 from eth_hash.auto import keccak
 from hexbytes import HexBytes
 
-from brownie._config import BROWNIE_FOLDER, CONFIG
+from brownie._config import CONFIG, DATA_FOLDER
 from brownie._singleton import _Singleton
 from brownie.convert import Wei, to_address
 from brownie.exceptions import IncompatibleEVMVersion, UnknownAccount, VirtualMachineError
@@ -100,7 +100,7 @@ class Accounts(metaclass=_Singleton):
 
         Returns:
             Account instance."""
-        project_path = BROWNIE_FOLDER.joinpath("data/accounts")
+        project_path = DATA_FOLDER.joinpath("accounts")
         if not filename:
             return [i.stem for i in project_path.glob("*.json")]
         filename = str(filename)
@@ -353,12 +353,12 @@ class LocalAccount(_PrivateKeyAccount):
 
         Attributes:
             filename: path to keystore file. If no folder is given, saved in
-                      brownie/data/accounts
+                      ~/.brownie/accounts
             overwrite: if True, will overwrite an existing file.
 
         Returns the absolute path to the keystore file as a string.
         """
-        path = BROWNIE_FOLDER.joinpath("data/accounts")
+        path = DATA_FOLDER.joinpath("accounts")
         path.mkdir(exist_ok=True)
         filename = str(filename)
         if not filename.endswith(".json"):